refactor: clean up routing structure and resolve duplicate path issue
