--- Client/src/config/axios.ts
@@ -1,42 +1,0 @@
-import axios from 'axios'
-import { API_URL } from './env'
-
-// Create axios instance with base URL
-const axiosInstance = axios.create({
-  baseURL: API_URL,
-  headers: {
-    'Content-Type': 'application/json'
-  }
-})
-
-// Add request interceptor for authentication
-axiosInstance.interceptors.request.use(
-  (config) => {
-    const userStr = localStorage.getItem('user')
-    if (userStr) {
-      const user = JSON.parse(userStr)
-      if (user && user.token) {
-        config.headers.Authorization = `Bearer ${user.token}`
-      }
-    }
-    return config
-  },
-  (error) => {
-    return Promise.reject(error)
-  }
-)
-
-// Add response interceptor for error handling
-axiosInstance.interceptors.response.use(
-  (response) => response,
-  (error) => {
-    // Handle 401 errors (unauthorized)
-    if (error.response && error.response.status === 401) {
-      localStorage.removeItem('user')
-      window.location.href = '/login'
-    }
-    return Promise.reject(error)
-  }
-)
-
-export default axiosInstance
