--- Client/src/config/theme.ts
@@ -1,1 +1,0 @@
-
