+++ Client/src/pages/About/sections/Hero/Hero.tsx
