+++ Client/src/pages/About/sections/Mission/Mission.tsx
