+++ Client/src/pages/About/sections/Team/Team.tsx
