+++ Client/src/pages/About/sections/Value/Value.tsx
