--- Client/src/pages/Bookings/Bookings.module.css
@@ -1,114 +1,0 @@
-.bookingsContainer {
-  padding: 2rem;
-  max-width: 1280px;
-  margin: 0 auto;
-  background: linear-gradient(135deg, #0f172a, #1e293b);
-  min-height: 100vh;
-  width: 100%;
-}
-
-.title {
-  font-size: 2.5rem;
-  font-weight: 700;
-  margin-bottom: 1rem;
-  background: linear-gradient(135deg, #00B4D8, #CAF0F8);
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-}
-
-.subtitle {
-  color: #CAF0F8;
-  font-size: 1.125rem;
-  margin-bottom: 2rem;
-}
-
-.bookingsGrid {
-  display: grid;
-  grid-template-columns: 1fr;
-  gap: 1.5rem;
-}
-
-@media (min-width: 768px) {
-  .bookingsGrid {
-    grid-template-columns: repeat(2, 1fr);
-  }
-}
-
-@media (min-width: 1024px) {
-  .bookingsGrid {
-    grid-template-columns: repeat(3, 1fr);
-  }
-}
-
-@media (max-width: 768px) {
-  .bookingsContainer {
-    padding: 1.5rem;
-  }
-}
-
-@media (max-width: 480px) {
-  .bookingsContainer {
-    padding: 1rem;
-  }
-}
-
-.bookingCard {
-  background: rgba(31, 41, 55, 0.5);
-  backdrop-filter: blur(10px);
-  border-radius: 0.5rem;
-  padding: 1.5rem;
-  border: 1px solid rgba(0, 180, 216, 0.1);
-  transition: all 0.3s ease;
-}
-
-.bookingCard:hover {
-  transform: translateY(-5px);
-  box-shadow: 0 10px 20px rgba(0, 180, 216, 0.2);
-  border-color: rgba(0, 180, 216, 0.3);
-}
-
-.bookingTitle {
-  font-size: 1.25rem;
-  font-weight: 600;
-  color: #00B4D8;
-  margin-bottom: 0.75rem;
-}
-
-.bookingDetails {
-  color: #CAF0F8;
-  font-size: 0.875rem;
-  line-height: 1.6;
-}
-
-.bookingActions {
-  display: flex;
-  gap: 1rem;
-  margin-top: 1rem;
-}
-
-.actionButton {
-  background: linear-gradient(135deg, #00B4D8, #0096C7);
-  color: white;
-  padding: 0.5rem 1rem;
-  border-radius: 0.375rem;
-  font-weight: 600;
-  font-size: 0.875rem;
-  border: none;
-  cursor: pointer;
-  transition: all 0.3s ease;
-}
-
-.actionButton:hover {
-  transform: translateY(-2px);
-  box-shadow: 0 5px 15px rgba(0, 180, 216, 0.3);
-}
-
-.cancelButton {
-  background: transparent;
-  color: #00B4D8;
-  border: 1px solid #00B4D8;
-}
-
-.cancelButton:hover {
-  background: rgba(0, 180, 216, 0.1);
-}
