--- Client/src/pages/Dashboard/Dashboard.module.css
@@ -1,73 +1,0 @@
-.dashboardContainer {
-  padding: 2rem;
-  max-width: 1280px;
-  margin: 0 auto;
-  background: linear-gradient(135deg, #0f172a, #1e293b);
-  min-height: 100vh;
-  width: 100%;
-}
-
-.heading {
-  font-size: 1.875rem;
-  margin-bottom: 1.5rem;
-  background: linear-gradient(135deg, #00B4D8, #CAF0F8);
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-}
-
-.dashboardContent {
-  display: grid;
-  grid-template-columns: 1fr;
-  gap: 1.5rem;
-}
-
-@media (min-width: 768px) {
-  .dashboardContent {
-    grid-template-columns: 1fr 1fr;
-  }
-}
-
-@media (min-width: 1024px) {
-  .dashboardContent {
-    grid-template-columns: 1fr 1fr 1fr;
-  }
-}
-
-.card {
-  background: rgba(31, 41, 55, 0.5);
-  backdrop-filter: blur(10px);
-  border-radius: 0.5rem;
-  box-shadow: 0 4px 6px rgba(0, 180, 216, 0.1);
-  padding: 1.5rem;
-  border: 1px solid rgba(0, 180, 216, 0.1);
-  transition: all 0.3s ease;
-}
-
-.card:hover {
-  transform: translateY(-5px);
-  box-shadow: 0 10px 20px rgba(0, 180, 216, 0.2);
-  border-color: rgba(0, 180, 216, 0.3);
-}
-
-.cardTitle {
-  font-size: 1.25rem;
-  font-weight: 600;
-  margin-bottom: 0.75rem;
-  color: #00B4D8;
-}
-
-.cardContent {
-  color: #CAF0F8;
-}
-
-@media (max-width: 768px) {
-  .dashboardContainer {
-    padding: 1.5rem;
-  }
-}
-
-@media (max-width: 480px) {
-  .dashboardContainer {
-    padding: 1rem;
-  }
-}
