+++ Client/src/pages/Home/Home.tsx
@@ -1,8 +1,8 @@
-import Hero from "./components/Hero/Hero";
-import Event from "./components/Event/Event";
-import Venue from "./components/Venue/Venue";
-import BookingGuide from "./components/BookingGuide/BookingGuide";
-import JoinUs from "./components/JoinUs/JoinUs";
+import Hero from "./sections/Hero/Hero";
+import Event from "./sections/Event/Event";
+import Venue from "./sections/Venue/Venue";
+import BookingGuide from "./sections/BookingGuide/BookingGuide";
+import JoinUs from "./sections/JoinUs/JoinUs";
 import styles from './Home.module.css';
 
 const Home = () => {
