--- Client/src/pages/Profile/Profile.module.css
@@ -1,164 +1,0 @@
-.profileContainer {
-  padding: 2rem;
-  max-width: 1280px;
-  margin: 0 auto;
-  background: linear-gradient(135deg, #0f172a, #1e293b);
-  min-height: 100vh;
-  width: 100%;
-}
-
-.title {
-  font-size: 2.5rem;
-  font-weight: 700;
-  margin-bottom: 1rem;
-  background: linear-gradient(135deg, #00B4D8, #CAF0F8);
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-}
-
-.subtitle {
-  color: #CAF0F8;
-  font-size: 1.125rem;
-  margin-bottom: 2rem;
-}
-
-.profileContent {
-  display: grid;
-  grid-template-columns: 1fr;
-  gap: 2rem;
-}
-
-@media (min-width: 768px) {
-  .profileContent {
-    grid-template-columns: 300px 1fr;
-  }
-}
-
-.profileSidebar {
-  background: rgba(31, 41, 55, 0.5);
-  backdrop-filter: blur(10px);
-  border-radius: 0.5rem;
-  padding: 1.5rem;
-  border: 1px solid rgba(0, 180, 216, 0.1);
-}
-
-.profileMain {
-  background: rgba(31, 41, 55, 0.5);
-  backdrop-filter: blur(10px);
-  border-radius: 0.5rem;
-  padding: 1.5rem;
-  border: 1px solid rgba(0, 180, 216, 0.1);
-}
-
-.profileAvatar {
-  width: 120px;
-  height: 120px;
-  border-radius: 50%;
-  margin: 0 auto 1rem;
-  border: 3px solid #00B4D8;
-  padding: 3px;
-}
-
-.profileInfo {
-  text-align: center;
-  margin-bottom: 2rem;
-}
-
-.profileName {
-  font-size: 1.5rem;
-  font-weight: 600;
-  color: #00B4D8;
-  margin-bottom: 0.5rem;
-}
-
-.profileEmail {
-  color: #CAF0F8;
-  font-size: 0.875rem;
-}
-
-.profileMenu {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-}
-
-.menuItem {
-  padding: 0.75rem 1rem;
-  color: #CAF0F8;
-  cursor: pointer;
-  transition: all 0.3s ease;
-  border-radius: 0.375rem;
-}
-
-.menuItem:hover {
-  background: rgba(0, 180, 216, 0.1);
-  color: #00B4D8;
-}
-
-.menuItem.active {
-  background: rgba(0, 180, 216, 0.2);
-  color: #00B4D8;
-}
-
-.sectionTitle {
-  font-size: 1.5rem;
-  font-weight: 600;
-  color: #00B4D8;
-  margin-bottom: 1.5rem;
-}
-
-.formGroup {
-  margin-bottom: 1.5rem;
-}
-
-.formLabel {
-  display: block;
-  color: #CAF0F8;
-  font-size: 0.875rem;
-  margin-bottom: 0.5rem;
-}
-
-.formInput {
-  width: 100%;
-  padding: 0.75rem;
-  border-radius: 0.375rem;
-  border: 1px solid rgba(0, 180, 216, 0.2);
-  background: rgba(31, 41, 55, 0.3);
-  color: #CAF0F8;
-  transition: all 0.3s ease;
-}
-
-.formInput:focus {
-  outline: none;
-  border-color: #00B4D8;
-  box-shadow: 0 0 0 2px rgba(0, 180, 216, 0.2);
-}
-
-.saveButton {
-  background: linear-gradient(135deg, #00B4D8, #0096C7);
-  color: white;
-  padding: 0.75rem 1.5rem;
-  border-radius: 0.375rem;
-  font-weight: 600;
-  font-size: 0.875rem;
-  border: none;
-  cursor: pointer;
-  transition: all 0.3s ease;
-}
-
-.saveButton:hover {
-  transform: translateY(-2px);
-  box-shadow: 0 5px 15px rgba(0, 180, 216, 0.3);
-}
-
-@media (max-width: 768px) {
-  .profileContainer {
-    padding: 1.5rem;
-  }
-}
-
-@media (max-width: 480px) {
-  .profileContainer {
-    padding: 1rem;
-  }
-}
