+++ Client/src/routes/privateRoutes.tsx
@@ -1,20 +1,6 @@
 import { RouteType } from '../types/routes'
-import Dashboard from '../pages/Dashboard/Dashboard'
-import Profile from '../pages/Profile/Profile'
-import Bookings from '../pages/Bookings/Bookings'
 
 export const privateRoutes: RouteType[] = [
   // Add private routes here (requiring authentication)
-  {
-    path: '/dashboard',
-    element: <Dashboard />
-  },
-  {
-    path: '/profile',
-    element: <Profile />
-  },
-  {
-    path: '/bookings',
-    element: <Bookings />
-  }
+
 ]
