--- Client/src/store/slices/authSlice.ts
@@ -1,50 +1,0 @@
-import { createSlice, PayloadAction } from '@reduxjs/toolkit'
-
-interface User {
-  id: string
-  name: string
-  email: string
-  token: string
-}
-
-interface AuthState {
-  user: User | null
-  isAuthenticated: boolean
-  isLoading: boolean
-  error: string | null
-}
-
-const initialState: AuthState = {
-  user: null,
-  isAuthenticated: false,
-  isLoading: false,
-  error: null
-}
-
-const authSlice = createSlice({
-  name: 'auth',
-  initialState,
-  reducers: {
-    loginStart: (state: AuthState) => {
-      state.isLoading = true
-      state.error = null
-    },
-    loginSuccess: (state: AuthState, action: PayloadAction<User>) => {
-      state.isLoading = false
-      state.isAuthenticated = true
-      state.user = action.payload
-    },
-    loginFailure: (state: AuthState, action: PayloadAction<string>) => {
-      state.isLoading = false
-      state.error = action.payload
-    },
-    logout: (state: AuthState) => {
-      state.user = null
-      state.isAuthenticated = false
-    }
-  }
-})
-
-export const { loginStart, loginSuccess, loginFailure, logout } = authSlice.actions
-
-export default authSlice.reducer
