--- Client/src/types/routes.ts
@@ -1,6 +1,0 @@
-import { ReactNode } from 'react'
-
-export interface RouteType {
-  path: string
-  element: ReactNode
-}
